finally fixed tests
