+++ tests/codegen/sample_workflows.py
@@ -399,34 +399,3 @@ def task_spark_jar():
         main_class_name="com.example.MainClass",
         parameters=["param1", "param2"],
     )
-
-
-wf_serverless = Workflow(
-    "brickflow-serverless-demo",
-    schedule_quartz_expression="0 0/20 0 ? * * *",
-    libraries=[PypiTaskLibrary(package="pytz==2024.2")],
-)
-
-
-@wf_serverless.task
-def entrypoint_task():
-    pass
-
-
-@wf_serverless.notebook_task
-def notebook_task():
-    return NotebookTask(
-        notebook_path="notebooks/example_notebook.py",
-        base_parameters={
-            "some_parameter": "some_value",  # in the notebook access these via dbutils.widgets.get("some_parameter")
-        },
-    )  # type: ignore
-
-
-@wf_serverless.spark_python_task
-def spark_python_task():
-    return SparkPythonTask(
-        python_file="/src/python/example.py",
-        source="GIT",
-        parameters=["--timezone", "UTC"],
-    )  # type: ignore
